Put configuration in config files so as to make it easier to see how a particular tool is configured. Also added collectCoverageFrom to get coverage from all appropriate modules even if no test directly tests them.
