+++ assets/buble.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+    objectAssign: 'Object.assign',
+    transforms: {
+        forOf: false,
+        generator: false
+    }
+};
